+++ test/Parser.js
@@ -134,6 +134,42 @@ describe('Parser', function() {
 			right: {type: 'Literal', value: 5}
 		});
 	});
+	it('should handle object literals', function() {
+		inst.addTokens(Lexer.tokenize('{foo: "bar", tek: 1+2}'));
+		inst.complete().should.deep.equal({
+			type: 'ObjectLiteral',
+			value: {
+				foo: {type: 'Literal', value: 'bar'},
+				tek: {
+					type: 'BinaryExpression',
+					operator: '+',
+					left: {type: 'Literal', value: 1},
+					right: {type: 'Literal', value: 2}
+				}
+			}
+		});
+	});
+	it('should handle nested object literals', function() {
+		inst.addTokens(Lexer.tokenize('{foo: {bar: "tek"}}'));
+		inst.complete().should.deep.equal({
+			type: 'ObjectLiteral',
+			value: {
+				foo: {
+					type: 'ObjectLiteral',
+					value: {
+						bar: {type: 'Literal', value: 'tek'}
+					}
+				}
+			}
+		});
+	});
+	it('should handle empty object literals', function() {
+		inst.addTokens(Lexer.tokenize('{}'));
+		inst.complete().should.deep.equal({
+			type: 'ObjectLiteral',
+			value: {}
+		});
+	});
 	it('should chain traversed identifiers', function() {
 		inst.addTokens(Lexer.tokenize('foo.bar.baz + 1'));
 		inst.complete().should.deep.equal({
@@ -155,20 +191,27 @@ describe('Parser', function() {
 		});
 	});
 	it('should apply transforms and arguments', function() {
-		inst.addTokens(Lexer.tokenize('foo|tr1|tr2.baz|tr3{bar:"tek"}'));
+		inst.addTokens(Lexer.tokenize('foo|tr1|tr2.baz|tr3({bar:"tek"})'));
 		inst.complete().should.deep.equal({
 			type: 'Transform',
 			name: 'tr3',
-			args: {bar: {type: 'Literal', value: 'tek'}},
+			args: [{
+				type: 'ObjectLiteral',
+				value: {
+					bar: {type: 'Literal', value: 'tek'}
+				}
+			}],
 			subject: {
 				type: 'Identifier',
 				value: 'baz',
 				from: {
 					type: 'Transform',
 					name: 'tr2',
+					args: [],
 					subject: {
 						type: 'Transform',
 						name: 'tr1',
+						args: [],
 						subject: {
 							type: 'Identifier',
 							value: 'foo'
@@ -176,6 +219,19 @@ describe('Parser', function() {
 					}
 				}
 			}
+		});
+	});
+	it('should handle multiple arguments in transforms', function() {
+		inst.addTokens(Lexer.tokenize('foo|bar("tek", 5, true)'));
+		inst.complete().should.deep.equal({
+			type: 'Transform',
+			name: 'bar',
+			args: [
+				{type: 'Literal', value: 'tek'},
+				{type: 'Literal', value: 5},
+				{type: 'Literal', value: true}
+			],
+			subject: {type: 'Identifier', value: 'foo'}
 		});
 	});
 	it('should apply filters to identifiers', function() {
@@ -207,6 +263,41 @@ describe('Parser', function() {
 					expr: {type: 'Literal', value: 1},
 					subject: {type: 'Identifier', value: 'foo'}
 				}
+			}
+		});
+	});
+	it('should allow dot notation for all operands', function() {
+		inst.addTokens(Lexer.tokenize('"foo".length + {foo: "bar"}.foo'));
+		inst.complete().should.deep.equal({
+			type: 'BinaryExpression',
+			operator: '+',
+			left: {
+				type: 'Identifier',
+				value: 'length',
+				from: {type: 'Literal', value: 'foo'}
+			},
+			right: {
+				type: 'Identifier',
+				value: 'foo',
+				from: {
+					type: 'ObjectLiteral',
+					value: {
+						foo: {type: 'Literal', value: 'bar'}
+					}
+				}
+			}
+		});
+	});
+	it('should allow dot notation on subexpressions', function() {
+		inst.addTokens(Lexer.tokenize('("foo" + "bar").length'));
+		inst.complete().should.deep.equal({
+			type: 'Identifier',
+			value: 'length',
+			from: {
+				type: 'BinaryExpression',
+				operator: '+',
+				left: {type: 'Literal', value: 'foo'},
+				right: {type: 'Literal', value: 'bar'}
 			}
 		});
 	});